Google runner project integration
removing event compare. Changing google runner to use the config to decide if it will run in test mode. It will also send emails if changes are detected and send_emails is true in the config. Sender is now chageable in the config; it still must match the appkey defined in .env.